Add i18n support
All content is stored in package translate.
For now only English is supported.

By keeping translations in Go structs naming conflicts, renaming,
auto-completion and so on make the system more stable to work with.

Since every message and payload sends back some kind of content,
each request needs to load the user's locale and appropriate content.
They are attached to a user struct, which is kind of a request scope.

Also translates Greeting.

Payloads also have been moved to a separate package payload.