Use bucket zeroscores to limit studies per day
To simplify the calulation of the offset used when calculating studytimes
we introduced a new bucket zeroscores.

The bucket is updated at every place where we change the score of a phrase,
but with the migration we update the bucket for all existing phrases.

Additionally, all studytimes are reset to times calculated with the new algorithm.